% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorVectorPickModule.R
\name{colorVectorPickModuleUI}
\alias{colorVectorPickModuleUI}
\alias{colorVectorPickModule}
\title{Shiny color vector editor}
\usage{
colorVectorPickModuleUI(id)

colorVectorPickModule(
  input,
  output,
  session,
  n_colors = reactive(NULL),
  current_colors = reactive(NULL),
  labels = reactive(NULL),
  show_order = reactive(NULL),
  palettes_from_pals = c(Parula = "parula", Blauw = "brewer.blues", `Blauw - Rood` =
    "coolwarm", `Rood - Blauw` = "warmcool", `Geel - Rood` = "brewer.ylorrd", Oranje =
    "brewer.oranges", Ocean = "ocean.haline", Cubic = "cubicl", `Ocean 2` =
    "ocean.phase", Viridis = "viridis", Rainbow = "kovesi.rainbow", Jet = "jet")
)
}
\arguments{
\item{id}{Shiny input ID}

\item{input}{Shiny session, do not set}

\item{n_colors}{Number of colors - reactive}

\item{current_colors}{Vector of current colors - reactive}

\item{labels}{Vector of labels - reactive}

\item{show_order}{Vector of ordering of the colors (integer) -  reactive}

\item{label}{Label for button}

\item{icon}{An icon for the button}

\item{status}{Button status color in \code{\link[softui:action_button]{softui::action_button()}}}
}
\description{
Shiny color vector editor
}
