% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlEditModule.R
\name{htmlEditModuleUI}
\alias{htmlEditModuleUI}
\alias{htmlEditModule}
\title{HTML edit a field (TinyMCE)}
\usage{
htmlEditModuleUI(id, label = "Edit HTML", ...)

htmlEditModule(input, output, session, title = "", value = reactive(""))
}
\description{
HTML edit a field (TinyMCE)
}
