% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlInput.R
\name{htmlInput}
\alias{htmlInput}
\alias{updatehtmlInput}
\title{Shiny HTML input field}
\usage{
htmlInput(
  inputId,
  label = NULL,
  value = NULL,
  height = 500,
  inline = FALSE,
  branding = FALSE,
  menubar = FALSE,
  toolbar =
    "styleselect | bold italic | numlist bullist | outdent indent | undo redo | insertdatetime",
  ...
)

updatehtmlInput(
  inputId,
  value,
  session = getDefaultReactiveDomain(),
  asis = FALSE
)
}
\arguments{
\item{inputId}{The Shiny input ID}

\item{value}{Initial value}

\item{height}{The height of the editor in pixels}

\item{inline}{If TRUE, the editor is displayed in an inline block}

\item{branding}{If TRUE, give some love to the TINY team}

\item{menubar}{If TRUE, displays a more advanced menu}

\item{toolbar}{The buttons to display above the editor.}
}
\description{
A rich text editor for Shiny applications with the TinyMCE editor.
}
\details{
Set options(mce_api_key = "MYKEY"). Get your free API key at \link{www.tiny.cloud}
}
