% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{from_json}
\alias{from_json}
\alias{to_json}
\title{Json Utilities}
\usage{
from_json(x, ...)

to_json(x, ...)
}
\description{
Wrapper around \code{jsonlite::fromJSON}, returns input argument if it
cannot be converted from JSON.
}
