% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonEditModule.R
\name{jsonEditModuleUI}
\alias{jsonEditModuleUI}
\alias{jsonEditModule}
\title{Shiny JSON edit module}
\usage{
jsonEditModuleUI(
  id,
  class = "",
  options = c("delete", "add"),
  options_labels = c(delete = "Laatste verwijderen", add = "Toevoegen", save =
    "Opslaan")
)

jsonEditModule(
  input,
  output,
  session,
  edit = c("key", "value"),
  widths = c(6, 6),
  value = reactive(NULL)
)
}
\arguments{
\item{class}{CSS class of the editor}

\item{opties}{"verwijderen", "toevoegen", or both}
}
\description{
Editor for key-value type JSON fields. Edit either or both of the keys and the labels.
}
