% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonOrderModule.R
\name{jsonOrderModuleUI}
\alias{jsonOrderModuleUI}
\alias{jsonOrderModule}
\title{Shiny JSON order edit module}
\usage{
jsonOrderModuleUI(id)

jsonOrderModule(
  input,
  output,
  session,
  data = NULL,
  order_column = NULL,
  label_column = NULL
)
}
\arguments{
\item{class}{CSS class of the editor}

\item{opties}{"verwijderen", "toevoegen", or both}
}
\description{
Editor for ordering a JSON vector e.g. [1,2,3,4]
}
