% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorVectorPickModule.R
\name{colorVectorPickModuleUI}
\alias{colorVectorPickModuleUI}
\alias{colorVectorPickModule}
\title{Shiny color vector editor}
\usage{
colorVectorPickModuleUI(
  id,
  label = "Kleuren kiezen",
  icon = softui::bsicon("palette-fill"),
  class = "btn-success"
)

colorVectorPickModule(
  input,
  output,
  session,
  n_colors = reactive(NULL),
  current_colors = reactive(NULL),
  labels = reactive(NULL),
  show_order = reactive(NULL),
  callback = function(data) { }
)
}
\description{
Shiny color vector editor
}
