% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorVectorPickModule.R
\name{colorVectorPickModuleUI}
\alias{colorVectorPickModuleUI}
\alias{colorVectorPickModule}
\title{Shiny color vector editor}
\usage{
colorVectorPickModuleUI(
  id,
  label = "Kleuren kiezen",
  icon = softui::bsicon("palette-fill"),
  status = "success"
)

colorVectorPickModule(
  input,
  output,
  session,
  n_colors = reactive(NULL),
  current_colors = reactive(NULL),
  labels = reactive(NULL),
  show_order = reactive(NULL),
  callback = function(data) { }
)
}
\arguments{
\item{id}{Shiny input ID}

\item{label}{Label for button}

\item{icon}{An icon for the button}

\item{status}{Button status color in \code{\link[softui:action_button]{softui::action_button()}}}

\item{input}{Shiny session, do not set}

\item{n_colors}{Number of colors - reactive}

\item{current_colors}{Vector of current colors - reactive}

\item{labels}{Vector of labels - reactive}

\item{show_order}{Vector of ordering of the colors (integer) -  reactive}

\item{callback}{Function to call after the colors have been chosen.}
}
\description{
Shiny color vector editor
}
