% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonMultiEdit.R
\name{jsonMultiEditUI}
\alias{jsonMultiEditUI}
\alias{jsonMultiEdit}
\title{JSON Multi-edit for a nested JSON list}
\usage{
jsonMultiEditUI(id, ui_header = NULL, choices_label = "Keuzelijst")

jsonMultiEdit(
  input,
  output,
  session,
  edit_names = FALSE,
  key = reactive("[]"),
  value = reactive("[]"),
  select_label = "Categorie",
  trigger = reactive(NULL),
  json = TRUE
)
}
\description{
Use case: choice of a selectinput might depend on what is chosen in another
selectinput. This nested dependency can be configured with jsonMultiEdit
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
