% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonOrderModule.R
\name{jsonOrderModuleUI}
\alias{jsonOrderModuleUI}
\alias{jsonOrderModule}
\title{Shiny JSON order edit module}
\usage{
jsonOrderModuleUI(id, label = "Change order", icon = NULL, class = "")

jsonOrderModule(
  input,
  output,
  session,
  data = reactive(NULL),
  order_column = reactive(NULL),
  label_column = reactive(NULL),
  callback = function(data) { }
)
}
\arguments{
\item{class}{CSS class of the editor}

\item{opties}{"verwijderen", "toevoegen", or both}
}
\description{
Editor for ordering a JSON vector e.g. \link{1,2,3,4}
}
