% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shintocatman-package.R
\docType{package}
\name{shintocatman-package}
\alias{shintocatman}
\alias{shintocatman-package}
\title{shintocatman: Category management}
\description{
Shiny modules and tools for category management
}
\keyword{internal}
